% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_years.R
\name{dog_years}
\alias{dog_years}
\title{Convert human years to dog years, using the 2019 formula published by UCSD School of Medicine.}
\usage{
dog_years(human_years)
}
\arguments{
\item{human_years}{Age in human years.}
}
\value{
Numeric dog equivalent age.
}
\description{
Convert human years to dog years, using the 2019 formula published by UCSD School of Medicine.
}
\examples{
dog_years(26)
}
