% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greet.R
\name{greet}
\alias{greet}
\title{Greet someone by name}
\usage{
greet(name, loud = FALSE)
}
\arguments{
\item{name}{A character string. The name to greet.}

\item{loud}{Logical. Should the greeting be uppercase? Default is FALSE.}
}
\value{
A character string containing the greeting.
}
\description{
Greet someone by name
}
\examples{
greet("America")
greet("America", loud = TRUE)
}
