% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greet_mean.R
\name{greet_mean}
\alias{greet_mean}
\title{Greet someone by name, another version of greet}
\usage{
greet_mean(name, loud = FALSE)
}
\arguments{
\item{name}{A character string. The name to greet.}

\item{loud}{Logical. Should the greeting be uppercase? Default is FALSE.}
}
\value{
A character string containing the greeting.
}
\description{
Greet someone by name, another version of greet
}
\examples{
greet_mean("America")
greet_mean("America", loud = TRUE)
}
