% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_lm.R
\name{quick_lm}
\alias{quick_lm}
\title{Fit and plot a simple linear regression}
\usage{
quick_lm(x, y)
}
\arguments{
\item{x}{Predictor (numeric vector).}

\item{y}{Response (numeric vector).}
}
\value{
The fitted lm object.
}
\description{
Fit and plot a simple linear regression
}
\examples{
predictor <- c(2,3,4,6,7,8,9,10)
response <- c(4,5,7,8,8,6,8,10)
quick_lm(predictor, response)
}
