% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greet.R
\name{greet}
\alias{greet}
\title{Greet the user with commentary based on the time!}
\usage{
greet(name = "researcher")
}
\arguments{
\item{name}{A character string. The name to greet.}
}
\value{
A character string containing the greeting with some funny options based on time of day.
}
\description{
Greet the user with commentary based on the time!
}
\examples{
greet(name = "America")
}
